README.md
This gun uses a system where the DLL is hosted on a webhost, ie the DLL is not inside the Loader. This is a system widely used in large loaders to protect the DLL from cheat in a private hosting. 

Loader has a game process detection system, as soon as the game is executed the cheat will be automatically injected.